Fix critical thread-safety bugs and optimize player lookups. Network robustness and protocol constant centralization
Critical Bug Fixes:

  - PacketHandler.cpp: Fix copy-paste bug where both direction and facing
    read from data[1], causing facing to always equal direction. Now
    correctly reads direction from data[1] and facing from data[2].

  - GameServer.cpp: Fix use-after-free in OnClientDisconnect by capturing
    player data (id, position) before removing from spatial hash and registry.

  - ClientConnection.cpp: Fix race condition in Disconnect() by moving all
    state cleanup (player removal, client removal, limiter update) to game
    thread via OnClientDisconnect queue. Prevents crashes when IO thread and
    game thread both try to modify player state.

  - ClientConnection.cpp: Fix double-disconnect race in FailHandshake() by
    using the same atomic disconnecting_ guard as Disconnect(). Prevents
    corrupted connection counts when handshake timeout fires while processing
    invalid packet.

  - Log.h: Make Log::Level atomic to prevent data races when log level is
    read from game thread and written from console thread.

  - BandwidthMonitor.h: Fix packet counter that was adding byte count instead
    of 1, causing massively inflated packet statistics.

  - ClientManager.cpp: Fix use-after-move in CloseAll() by using swap instead
    of move, ensuring snapshot contains valid data for iteration.

  Major Improvements:

  - Protocol.h: Remove "using std::vector" directive that polluted global
    namespace for all including files.

  - Protocol.h: Fix potential integer overflow in PacketReader bounds checks
    by using subtraction instead of addition.

  - PlayerRegistry.h: Add duplicate client_id check in CreatePlayer() to
    prevent orphaned players and broken lookups if same client logs in twice.

  - PlayerRegistry.h: Add reverse mapping (player_id -> client_id) for O(1)
    removal by player_id instead of O(N) linear search.

  - GameServer.cpp: Optimize BroadcastDirtyPlayers to store viewer pointer
    when building update map, avoiding redundant GetByClientID lookup.

  - GameServer.cpp: Remove ~100 lines of dead commented-out code.

  - LuaEngine.h/cpp: Change player_id parameter from uint32_t to uint64_t
    for consistency with rest of codebase.

  - ClientConnection.h: Add thread safety documentation to PingTracker.

  Performance Impact:

  - Player removal: O(N) -> O(1) via reverse lookup map
  - Dirty player broadcasting: Eliminated redundant client lookups per viewer

=============================================================

TCP Fragmentation Fix (Critical)
  - NetworkConnection.cs: Added ReadExact() helper method to handle TCP stream
    fragmentation. TCP may return fewer bytes than requested in a single Read()
    call, causing packet parsing failures. Now loops until all expected bytes
    are received or connection closes.

  PacketReader Bounds Checking
  - PacketReader.cs: Added CheckBounds() validation to all Read methods (ReadU8,
    ReadU16, ReadU32, ReadU64). Now throws descriptive ArgumentOutOfRangeException
    instead of IndexOutOfRangeException, with details like:
    "PacketReader.ReadU8: not enough bytes (need 1, have 0 at offset 12)"
  - HasBytes() now also checks for null data and negative offset.

  PacketWriter String Safety
  - PacketWriter.cs: Added warning and proper truncation for strings exceeding
    255 bytes (1-byte length prefix limit). Previously silently truncated via
    cast to byte.

  Protocol Constants Centralized
  - PacketHeader.cs: Consolidated all protocol constants in one place:
    - MaxIncomingPayload = 4096 (Server to Client limit)
    - MaxOutgoingPayload = 4096 (Client to Server limit)
    - Removed unused MaxProtocolPayload (was 65535)
  - NetworkConnection.cs: Now uses PacketHeader.MaxIncomingPayload
  - PacketWriterTests.cs: Now references PacketHeader.HeaderSize instead of
    duplicating the constant

  Code Cleanup
  - PlayerView.cs: Removed unused factory parameter from
    InitializeAsRemotePlayer(ulong playerId) - was passed but never used
  - PlayerViewFactory.cs: Updated call site to match new signature

+++ dyewars_server/src/lua/LuaEngine.cpp
@@ -18,7 +18,7 @@ LuaGameEngine::~LuaGameEngine() {
     }
 }
 
-void LuaGameEngine::OnPlayerMoved(uint32_t player_id, int x, int y, uint8_t direction) {
+void LuaGameEngine::OnPlayerMoved(uint64_t player_id, int x, int y, uint8_t direction) {
     // 1. Lock the Mutex (Thread Safety)
     std::lock_guard<std::mutex> lock(lua_mutex_);
